joyboard_game/requirements.txt
colorama==0.4.6
pygame==2.6.1
pyserial==3.5
Requests==2.32.4